requirements.txt
streamlit
tensorflow
pillow
gdown
pandas
folium
streamlit-folium
